refact: common components
A  src/app/(common)/notifications/_components/NotificationsGrid.jsx
A  src/app/(common)/notifications/_constants/columnsConstant.jsx
A  src/app/(common)/notifications/_hooks/notificationsHook.jsx
R  src/components/main-layout/app/ProfileGrid.jsx -> src/app/(common)/profile/_components/ProfileGrid.jsx
A  src/app/(common)/profile/_hooks/profileHook.jsx
M  src/app/doctoralCenter/admin/_hooks/eventManagementHook.jsx
M  src/app/doctoralCenter/admin/_hooks/homeHook.jsx
M  src/app/doctoralCenter/admin/_hooks/userManagementHook.jsx
M  src/app/doctoralCenter/admin/notifications/page.jsx
M  src/app/doctoralCenter/admin/profile/page.jsx
M  src/app/doctoralCenter/expert/_components/DoctralCenterHomeGrid.jsx
M  src/app/doctoralCenter/expert/_lib/UnauthorizedUsersGridData.jsx
D  src/components/main-layout/app/NotificationsGrid.jsx
M  src/components/main-layout/common/Header.jsx
