+++ src/app/(common)/profile/_components/ProfileGrid.jsx
@@ -12,98 +12,30 @@ import {
   Tab,
   Tooltip
 } from "@mui/material";
-import { useState } from "react";
 import { Logout } from "@mui/icons-material";
 import { TabContext, TabList, TabPanel } from "@mui/lab";
 import { InsertPhoto } from "@mui/icons-material";
-import Auth from "@/lib/auth/auth";
-import FileAPI from "@/api/file";
 import { HighlightOff } from "@mui/icons-material";
-import { setAlertBox } from "@/features/uiState/slices/uiStateSlice";
-import APIWrapper from "@/helpers/APIWrapper";
-import { useAppDispatch } from "@/features/constants";
 import CustomTable from "@/common/CustomTable";
-import { createDataUrl } from "@/helpers/utils";
 import Loading from "@/app/loading";
+import { ProfileHook } from "../_hooks/profileHook";
 
-export default function ProfileDataGrid({ user, nameFields, setUser }) {
-  const [isImageLoading, setIsImageLoading] = useState(false);
-  const { logAlert } = APIWrapper();
-  const { handleLogout } = Auth();
-  const dispatch = useAppDispatch();
-
-  const [deletePictureDialog, setDeletePictureDialog] = useState(false);
-  const { upload, deleteFile } = FileAPI();
-
-  const uploadPicture = async (event) => {
-    setIsImageLoading(true);
-    const fileBytes = event.target.files[0];
-    const { name, type } = fileBytes;
-
-    const formData = new FormData();
-    formData.append("data", fileBytes);
-    formData.append("filename", name);
-    formData.append("mimetype", type);
-
-    const result = await upload(formData, "avatar");
-    if (result != []) {
-      user.picture = result.data;
-      user.pictureBlob = await createDataUrl({
-        picture: fileBytes,
-        fileType: "file"
-      });
-      dispatch(setUser({ data: user }));
-
-      logAlert({
-        message: `Потребител ${user.email} си смени снимката`,
-        description: `Потребителят си смени снимката успешно!`,
-        action: "Потребител смени снимка",
-        level: "success"
-      });
-    } else {
-      dispatch(
-        setAlertBox({
-          message: "Грешка при качването на снимката!",
-          severity: "error"
-        })
-      );
-    }
-
-    setIsImageLoading(false);
-  };
-
-  const deletePicture = async () => {
-    setIsImageLoading(true);
-
-    const result = await deleteFile({ filename: user.picture }, "avatar");
-    if (result != []) {
-      user.picture = "";
-      user.pictureBlob = "";
-      dispatch(setUser({ data: user }));
-
-      logAlert({
-        message: `Потребител ${user.email} си изтри снимката`,
-        description: `Потребителят си изтри снимката успешно!`,
-        action: "Потребител изтри снимка",
-        level: "success"
-      });
-    } else {
-      dispatch(
-        setAlertBox({
-          message: "Грешка при изтриването на снимката!",
-          severity: "error"
-        })
-      );
-    }
-
-    setIsImageLoading(false);
-  };
+export default function ProfileGrid({ user, nameFields, setUser }) {
+  const {
+    uploadPicture,
+    handleLogout,
+    isImageLoading,
+    deletePictureDialog,
+    setDeletePictureDialog,
+    deletePicture
+  } = ProfileHook(setUser);
 
   return (
     <Box sx={{ width: "100%", maxWidth: { sm: "100%", md: "1700px" } }}>
       <Typography component="h2" variant="h6" sx={{ mb: 2 }}>
         {user.role}
       </Typography>
+
       <Stack
         direction="row"
         useFlexGap
