+++ src/app/doctoralCenter/admin/_hooks/userManagementHook.jsx
@@ -28,7 +28,7 @@ export const UserManagementHook = () => {
 
   useEffect(() => {
     fetchAuthorizedUsers();
-    runPeriodically(() => {
+    return runPeriodically(() => {
       fetchAuthorizedUsers();
     });
   }, [fetchAuthorizedUsers]);
