+++ src/app/doctoralCenter/admin/notifications/page.jsx
@@ -1,16 +1,16 @@
 "use client";
-import NotificationGrid from "@/components/main-layout/app/NotificationsGrid";
 import Layout from "@/components/main-layout/Layout";
 
 import { path } from "../_constants/pathConstant";
 import { sideMenu } from "../_constants/sideMenuConstants";
+import NotificationsGrid from "@/app/(common)/notifications/_components/NotificationsGrid";
 
 export default function Page() {
   return (
     <Layout
       headerTitle={"Известия"}
       basePath={path}
-      MainView={<NotificationGrid />}
+      MainView={<NotificationsGrid />}
       mainListItems={sideMenu}
     />
   );
