+++ src/app/doctoralCenter/expert/_lib/UnauthorizedUsersGridData.jsx
@@ -17,7 +17,7 @@ export default function UnauthorizedUsersGridData() {
     };
 
     getUnauthorizedUsers();
-    runPeriodically(() => {
+    return runPeriodically(() => {
       getUnauthorizedUsers();
     });
   }, [setRows]);
